@@ -4,41 +4,43 @@ source 'https://rails-assets.org' do
   gem 'rails-assets-tether', '>= 1.1.0'
 end
 
-ruby '2.3.0'
+ruby '3.1.2'
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 4.2.7.1'
+gem 'rails', '~> 7.0.2.3'
 # Use sqlite3 as the database for Active Record
 
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
+gem 'sass-rails', '~> 6.0'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
+gem 'coffee-rails', '~> 5.0.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
 
 # Use jquery as the JavaScript library
-gem 'jquery-rails'
+gem 'jquery-rails', '~> 4.4.0'
 # Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
 gem 'turbolinks'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
+gem 'jbuilder', '~> 2.11.5'
 # bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
+gem 'sdoc', '~> 2.3.1', group: :doc
+gem 'activesupport', '~> 7.0.2.3'
 gem 'API_Fuzzer', git: "https://github.com/lalithr95/API-fuzzer.git"
 #git: "https://github.com/lalithr95/API-fuzzer.git"
-gem 'highcharts-rails'
-gem 'newrelic_rpm'
-gem 'sidekiq'
+gem 'highcharts-rails', '~> 6.0.3'
+gem 'newrelic_rpm', '~> 8.6.0'
+gem 'sidekiq', '~> 6.4.1'
 gem 'bootstrap', '~> 4.0.0.alpha3.1'
-gem 'puma'
-gem 'pg'
-gem 'redis'
-gem 'sinatra', :require => false
-gem 'devise'
-gem 'rollbar'
-gem 'devise-bootstrap-views'
+gem 'puma', '~> 5.6.4'
+gem 'pg', '~> 1.3.5'
+gem 'redis', '~> 4.6.0'
+gem 'sinatra', '~> 2.2.0'
+gem 'devise', '~> 4.8.1'
+gem 'rollbar', '~> 3.3.0'
+gem 'devise-bootstrap-views', '~> 1.1.0'
+gem 'json', '~> 2.3.1'
 
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
@@ -50,15 +52,15 @@ gem 'devise-bootstrap-views'
 # gem 'capistrano-rails', group: :development
 
 group :development, :test do
-  gem 'sqlite3'
+  gem 'sqlite3', '~> 1.4.2'
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug'
 end
 
 group :development do
   # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
+  gem 'web-console', '~> 4.2.0'
 
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
-  gem 'spring'
+  gem 'spring', '~> 4.0.0'
 end
