@@ -1,4 +1,4 @@
-class CreateScans < ActiveRecord::Migration
+class CreateScans < ActiveRecord::Migration[7.0]
   def change
     create_table :scans do |t|
       t.string :url
