@@ -1,4 +1,4 @@
-class ChangeVulnerabilityTypeToClass < ActiveRecord::Migration
+class ChangeVulnerabilityTypeToClass < ActiveRecord::Migration[7.0]
   def change
     rename_column :vulnerabilities, :type, :class_type
   end
