@@ -1,4 +1,4 @@
-class AddJsonColumnToScans < ActiveRecord::Migration
+class AddJsonColumnToScans < ActiveRecord::Migration[7.0]
   def change
     add_column :scans, :json, :boolean, default: false
   end
