@@ -1,4 +1,4 @@
-class AddUserIdColumnToScans < ActiveRecord::Migration
+class AddUserIdColumnToScans < ActiveRecord::Migration[7.0]
   def change
     add_reference :scans, :user, index: true, foreign_key: true
   end
