@@ -1,4 +1,4 @@
-class AddStatusToScans < ActiveRecord::Migration
+class AddStatusToScans < ActiveRecord::Migration[7.0]
   def up
     add_column :scans, :status, :string, default: 'processing'
   end
