@@ -21,7 +21,7 @@ module FuzzerApp
     # config.i18n.default_locale = :de
 
     # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
+    # config.active_record.raise_in_transactional_callbacks = true
     config.autoload_paths += %w(#{config.root}/app/workers)
   end
 end
