@@ -1,60 +1,57 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20161014024647) do
-
+ActiveRecord::Schema[7.0].define(version: 2016_10_14_024647) do
   create_table "scans", force: :cascade do |t|
-    t.string   "url"
-    t.string   "sid"
-    t.string   "parameters"
-    t.string   "method"
-    t.string   "cookies"
-    t.datetime "created_at",                        null: false
-    t.datetime "updated_at",                        null: false
-    t.boolean  "json",       default: false
-    t.integer  "user_id"
-    t.string   "status",     default: "processing"
+    t.string "url"
+    t.string "sid"
+    t.string "parameters"
+    t.string "method"
+    t.string "cookies"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.boolean "json", default: false
+    t.integer "user_id"
+    t.string "status", default: "processing"
+    t.index ["user_id"], name: "index_scans_on_user_id"
   end
-
-  add_index "scans", ["user_id"], name: "index_scans_on_user_id"
 
   create_table "users", force: :cascade do |t|
-    t.string   "email",                  default: "", null: false
-    t.string   "encrypted_password",     default: "", null: false
-    t.string   "reset_password_token"
+    t.string "email", default: "", null: false
+    t.string "encrypted_password", default: "", null: false
+    t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          default: 0,  null: false
+    t.integer "sign_in_count", default: 0, null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
-    t.string   "current_sign_in_ip"
-    t.string   "last_sign_in_ip"
-    t.datetime "created_at",                          null: false
-    t.datetime "updated_at",                          null: false
+    t.string "current_sign_in_ip"
+    t.string "last_sign_in_ip"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["email"], name: "index_users_on_email", unique: true
+    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
-
-  add_index "users", ["email"], name: "index_users_on_email", unique: true
-  add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
 
   create_table "vulnerabilities", force: :cascade do |t|
-    t.string   "class_type"
-    t.integer  "scan_id"
-    t.string   "parameter"
-    t.text     "description"
-    t.string   "value"
-    t.datetime "created_at",  null: false
-    t.datetime "updated_at",  null: false
-    t.string   "status"
+    t.string "class_type"
+    t.integer "scan_id"
+    t.string "parameter"
+    t.text "description"
+    t.string "value"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "status"
   end
 
+  add_foreign_key "scans", "users"
 end
